Refactoring: moved all models into their own folder, ready to extend them to the network, classifier, regressor structure already used for CNN and MLP.
