+++ sktime_dl/deeplearning/__init__.py
@@ -1,8 +1,12 @@
-from ._encoder import EncoderClassifier
-from ._fcn import FCNClassifier
-from ._mcdcnn import MCDCNNClassifier
-from ._mcnn import MCNNClassifier
-from ._resnet import ResNetClassifier
-from ._tlenet import TLENETClassifier
-from ._twiesn import TWIESNClassifier
-from ._inceptiontime import InceptionTimeClassifier
+from .cnn._classifier import CNNClassifier
+from .cnn._regressor import CNNRegressor
+from .encoder._classifier import EncoderClassifier
+from .fcn._classifier import FCNClassifier
+from .mcdcnn._classifier import MCDCNNClassifier
+from .mcnn._classifier import MCNNClassifier
+from .mlp._classifier import MLPClassifier
+from .mlp._regressor import MLPRegressor
+from .resnet._classifier import ResNetClassifier
+from .tlenet._classifier import TLENETClassifier
+from .twiesn._classifier import TWIESNClassifier
+from .inceptiontime._classifier import InceptionTimeClassifier
