+++ sktime_dl/deeplearning/cnn/__init__.py
@@ -1,2 +1,1 @@
-from ._classifier import CNNClassifier
-from ._regressor import CNNRegressor
+
