+++ sktime_dl/deeplearning/encoder/_classifier.py
