+++ sktime_dl/deeplearning/mcdcnn/_classifier.py
