+++ sktime_dl/deeplearning/mcnn/_classifier.py
