+++ sktime_dl/deeplearning/mlp/__init__.py
@@ -1,2 +1,1 @@
-from ._classifier import MLPClassifier
-from ._regressor import MLPRegressor
+
