+++ sktime_dl/meta/_dltuner.py
@@ -3,7 +3,7 @@ __author__ = "James Large"
 import numpy as np
 
 from sktime_dl.base.estimators._classifier import BaseDeepClassifier
-from sktime_dl.deeplearning.cnn import CNNClassifier
+from sktime_dl.deeplearning import CNNClassifier
 
 from sklearn.model_selection import GridSearchCV
 from sklearn.model_selection import RandomizedSearchCV
